UAS/UAS.UnitTests/UAS.UnitTests/FrameworkUAD/DataAccess/IssueSplitFilterDetailsTest.cs
using System;
using System.Collections.Generic;
using System.Data;
using System.Data.SqlClient;
using System.Data.SqlClient.Fakes;
using System.Diagnostics.CodeAnalysis;
using System.Linq;
using FrameworkUAD.DataAccess;
using FrameworkUAD.DataAccess.Fakes;
using Microsoft.QualityTools.Testing.Fakes;
using NUnit.Framework;
using Shouldly;
using UAS.UnitTests.Helpers;
using Entity = FrameworkUAD.Entity;
using KMFakes = KM.Common.Fakes;
using ClientConnections = KMPlatform.Object.ClientConnections;

namespace UAS.UnitTests.FrameworkUAD.DataAccess
{
    /// <summary>
    /// Unit Tests for <see cref="IssueSplitFilterDetails"/>
    /// </summary>
    [TestFixture]
    [ExcludeFromCodeCoverage]
    public class IssueSplitFilterDetailsTest
    {
        private const int FilterId = 0;
        private const string ProcSelectFilterId = "e_IssueSplitFilterDetails_Select_FilterID";
        private const string MethodSelectFilterId = "SelectFilterID";
        private static readonly ClientConnections Client = new ClientConnections();

        private IDisposable _context;
        private IList<Entity.IssueSplitFilterDetails> _list;
        private Entity.IssueSplitFilterDetails _objWithRandomValues;
        private Entity.IssueSplitFilterDetails _objWithDefaultValues;
        private SqlCommand _sqlCommand;

        [SetUp]
        public void SetUp()
        {
            _context = ShimsContext.Create();

            _objWithRandomValues = typeof(Entity.IssueSplitFilterDetails).CreateInstance();
            _objWithDefaultValues = typeof(Entity.IssueSplitFilterDetails).CreateInstance(true);

            _list = new List<Entity.IssueSplitFilterDetails>
            {
                _objWithRandomValues,
                _objWithDefaultValues
            };
            SetupFakes();
        }

        [TearDown]
        public void TearDown()
        {
            _context.Dispose();
        }

        [Test]
        public void SelectFilterID_WhenCalled_VerifySqlParameters()
        {
            // Arrange, Act
            var result = typeof(IssueSplitFilterDetails).CallMethod(MethodSelectFilterId, new object[] { FilterId, Client }) as List<Entity.IssueSplitFilterDetails>;

            // Assert
            _sqlCommand.ShouldSatisfyAllConditions(
                () => _sqlCommand.Parameters["@FilterID"].Value.ShouldBe(FilterId),
                () => result.ShouldNotBeNull(),
                () => result.IsListContentMatched(_list.ToList()).ShouldBeTrue(),
                () => _sqlCommand.CommandText.ShouldBe(ProcSelectFilterId),
                () => _sqlCommand.CommandType.ShouldBe(CommandType.StoredProcedure));
        }

        private void SetupFakes()
        {
            ShimDataFunctions.GetClientSqlConnectionClientConnections = _ => new ShimSqlConnection().Instance;
            KMFakes.ShimDataFunctions.ExecuteNonQuerySqlCommand = cmd =>
            {
                _sqlCommand = cmd;
                return true;
            };

            ShimSqlCommand.AllInstances.ConnectionGet = cmd => new ShimSqlConnection().Instance;
            ShimDataFunctions.ExecuteReaderSqlCommand = cmd =>
            {
                _sqlCommand = cmd;
                return _list.GetSqlDataReader();
            };
        }
    }
}
